refactor : 시설안에 존재하는 timeslot에 관한 클래스를 패키지로 분리
