+++ src/main/java/com/example/rentalSystem/domain/facility/entity/timeTable/TimeStatus.java
@@ -1,6 +1,5 @@
-package com.example.rentalSystem.domain.facility.entity;
+package com.example.rentalSystem.domain.facility.entity.timeTable;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -8,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum RentalStatus {
+public enum TimeStatus {
     AVAILABLE("예약가능"), UNAVAILABLE("이용불가"), RESERVED("예약완료"), WAITING("예약대기"), CURRENT("현재예약");
     final String description;
 
@@ -16,5 +15,5 @@ public enum RentalStatus {
     public String getDescription() {
         return description;
     }
-    
+
 }
